+++ src/frames.h
@@ -1,13 +1,14 @@
 #ifndef HAL_MM_FRAMES_H
 #define HAL_MM_FRAMES_H
 
+#include <stddef.h>
 #include "placement_allocator.h"
 #include "paging.h"
 
-void dmm_set_frame(unsigned int addr);
-void dmm_clear_frame(unsigned int addr);
-unsigned int dmm_test_frame(unsigned int addr);
-unsigned int dmm_first_frame();
+void dmm_set_frame(size_t addr);
+void dmm_clear_frame(size_t addr);
+size_t dmm_test_frame(size_t addr);
+size_t dmm_first_frame();
 int dmm_alloc_frame(PageTableEntry *page, int is_kernel, int is_writable);
 void dmm_free_frame(PageTableEntry *page);
 void dmm_frames_init();
