+++ src/placement_allocator.c
@@ -1,11 +1,12 @@
 #include <dmm.h>
+#include <stddef.h>
 #include "placement_allocator.h"
 
-unsigned int placement;
+size_t placement;
 
-unsigned int dmm_pa_kernel_end;
+size_t dmm_pa_kernel_end;
 
-void dmm_placement_allocator_init(unsigned int kernel_end)
+void dmm_placement_allocator_init(size_t kernel_end)
 {
     dmm_pa_kernel_end = kernel_end;
     placement = dmm_pa_kernel_end;
@@ -15,7 +16,7 @@ void dmm_placement_allocator_init(unsigned int kernel_end)
  * Simple placement based allocator. This allocates off the end of the kernel.
  * Potential bugs: This has no bound, and does not check with the final heap allocator
  */
-void *kmalloc_int(unsigned int size, unsigned int flags)
+void *kmalloc_int(size_t size, size_t flags)
 {
     // Otherwise, increment to next placement address.
     placement += size;
