+++ src/placement_allocator.h
@@ -3,7 +3,9 @@
 
 #define MALLOC_ALIGN 1
 
-void dmm_placement_allocator_init(unsigned int kernel_end);
-void *kmalloc_int(unsigned int size, unsigned int flags);
+#include <stddef.h>
+
+void dmm_placement_allocator_init(size_t kernel_end);
+void *kmalloc_int(size_t size, size_t flags);
 
 #endif /* HAL_MM_PLACEMENT_ALLOCATOR */
